requirements.txt
streamlit
pymysql
python-dotenv
streamlit-option-menu
pillow
bcrypt
extra-streamlit-components
pytest 
pytest-mock 
httpx
